Finish segregating patmat cps hacks.
